bug fixes, compiler literal and name caching, basic bytecode generation

+++ src/runtime/values/complex.rs
@@ -5,12 +5,12 @@ use crate::runtime::values::RuntimeValue;
 
 #[derive(Eq, PartialEq, Clone, Debug)]
 pub struct ObjectValue {
-    pub id: u64,
-    pub properties: indexmap::IndexMap<RuntimeValue, u64>, // u64 is memory address
+    pub id: usize,
+    pub properties: indexmap::IndexMap<RuntimeValue, usize>, // usize is memory address
 }
 
 impl ObjectValue {
-    pub fn get_property(&self, key: &RuntimeValue) -> Result<Option<&u64>, AccessError> {
+    pub fn get_property(&self, key: &RuntimeValue) -> Result<Option<&usize>, AccessError> {
         match key {
             RuntimeValue::Int(_)
             | RuntimeValue::Float(_)
@@ -20,13 +20,13 @@ impl ObjectValue {
         }
     }
 
-    pub fn set_property(&mut self, key: &RuntimeValue, value: u64) {
+    pub fn set_property(&mut self, key: &RuntimeValue, value: usize) {
         self.properties.insert(key.clone(), value);
     }
 }
 
-impl From<indexmap::IndexMap<RuntimeValue, u64>> for ObjectValue {
-    fn from(properties: indexmap::IndexMap<RuntimeValue, u64>) -> Self {
+impl From<indexmap::IndexMap<RuntimeValue, usize>> for ObjectValue {
+    fn from(properties: indexmap::IndexMap<RuntimeValue, usize>) -> Self {
         Self {
             properties,
             id: Counter.next(),
@@ -44,26 +44,26 @@ impl Default for ObjectValue {
 }
 
 impl HasId for ObjectValue {
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
 
 #[derive(Eq, PartialEq, Clone, Debug)]
 pub struct ListValue {
-    pub id: u64,
-    pub elements: Vec<u64>, // u64 is memory address
+    pub id: usize,
+    pub elements: Vec<usize>, // usize is memory address
 }
 
 impl ListValue {
-    pub fn get_element(&self, index: &RuntimeValue) -> Result<Option<&u64>, AccessError> {
+    pub fn get_element(&self, index: &RuntimeValue) -> Result<Option<&usize>, AccessError> {
         match index {
             RuntimeValue::Int(i) => Ok(self.elements.get(i.value as usize)),
             _ => Err(AccessError::InvalidAddress),
         }
     }
 
-    pub fn set_element(&mut self, index: &RuntimeValue, value: u64) -> Result<(), AccessError> {
+    pub fn set_element(&mut self, index: &RuntimeValue, value: usize) -> Result<(), AccessError> {
         match index {
             RuntimeValue::Int(i) => Ok(self.elements[i.value as usize] = value),
             _ => Err(AccessError::InvalidAddress),
@@ -81,31 +81,31 @@ impl Default for ListValue {
 }
 
 impl HasId for ListValue {
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
 
 #[derive(Eq, PartialEq, Clone, Debug)]
 pub struct IdentityValue {
-    pub id: u64,
+    pub id: usize,
     pub address: Vec<RuntimeValue>,
 }
 
 impl HasId for IdentityValue {
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
 
 #[derive(Eq, PartialEq, Clone, Debug)]
 pub struct ReferenceValue {
-    pub id: u64,
-    pub memory_address: u64,
+    pub id: usize,
+    pub memory_address: usize,
 }
 
 impl HasId for ReferenceValue {
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
